requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
anthropic==0.18.1
pinecone-client==5.0.1
langchain==0.1.6
langchain-community==0.0.20
openai==1.12.0
pypdf==4.0.1
python-docx==1.1.0
python-pptx==0.6.23
pandas==2.2.0
numpy==1.26.3
pydantic==2.5.3
python-multipart==0.0.6
aiofiles==23.2.1
networkx==3.2.1
tiktoken==0.5.2
scikit-learn==1.4.0
sentence-transformers==2.3.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
supabase>=2.0.0
ffmpeg-python==0.2.0
moviepy==1.0.3